+++ src/localisation/english.ts
@@ -356,6 +356,7 @@ const ENGLISH: Translations = {
   [Phrase.UpdateAvailableRemindButtonText]: 'Remind Me Later',
   [Phrase.Saving]: 'Saving...',
   [Phrase.StartTyping]: 'Start typing...',
+  [Phrase.ToggleDrawingMode]: 'Toggle drawing mode',
   [Phrase.StarSelected]: 'Lock selected rows. Locked rows will not be automatically aged out.',
   [Phrase.UnstarSelected]: 'Unlock selected rows. Unlocked rows may be automatically aged out to make space for new recordings.',
   [Phrase.Selection]: 'Selection',