+++ src/localisation/german.ts
@@ -356,6 +356,7 @@ const GERMAN: Translations = {
   [Phrase.UpdateAvailableRemindButtonText]: 'Erinnere mich später',
   [Phrase.Saving]: 'Sparen...',
   [Phrase.StartTyping]: 'Beginnen Sie mit der Eingabe...',
+  [Phrase.ToggleDrawingMode]: 'Zeichenmodus umschalten',
   [Phrase.StarSelected]: 'Ausgewählte Zeilen sperren.',
   [Phrase.UnstarSelected]: 'Ausgewählte Zeilen entsperren.',
   [Phrase.Selection]: 'Auswahl',